+++ src/data/index.ts
@@ -1,2 +1,3 @@
 export * from './providers/balancer-api';
+export * from './providers/initPoolDataProvider';
 export * from './types';