+++ src/entities/initPool/index.ts
@@ -13,7 +13,7 @@ export class InitPool {
 
     buildCall(input: InitPoolInput, poolState: PoolState): InitPoolBuildOutput {
         this.inputValidator.validateAddLiquidity(input, poolState);
-        switch (poolState.balancerVersion) {
+        switch (poolState.vaultVersion) {
             case 2:
                 return new InitPoolV2().buildCall(input, poolState);
             case 3: