+++ src/entities/removeLiquidity/types.ts
@@ -67,7 +67,7 @@ export type RemoveLiquidityBaseQueryOutput = {
     amountsOut: TokenAmount[];
     tokenOutIndex?: number;
     toInternalBalance: boolean;
-    balancerVersion: 2 | 3;
+    vaultVersion: 2 | 3;
 };
 
 export type RemoveLiquidityComposableStableQueryOutput =