+++ test/lib/utils/swapHelpers.ts
@@ -37,7 +37,7 @@ export async function assertSwapExactIn(
     };
     let contractToCall = BALANCER_ROUTER[chainId];
 
-    if (swap.balancerVersion === 2) {
+    if (swap.vaultVersion === 2) {
         buildCallInput = {
             ...buildCallInput,
             sender: testAddress,
@@ -118,7 +118,7 @@ export async function assertSwapExactOut(
     };
     let contractToCall = BALANCER_ROUTER[chainId];
 
-    if (swap.balancerVersion === 2) {
+    if (swap.vaultVersion === 2) {
         buildCallInput = {
             ...buildCallInput,
             sender: testAddress,