+++ test/v2/addLiquidity/gyroEV2.integration.test.ts
@@ -222,7 +222,7 @@ class MockApi {
                     index: 1,
                 },
             ],
-            balancerVersion: 2,
+            vaultVersion: 2,
         };
     }
 }